+++ src/main/java/de/bixilon/minosoft/gui/rendering/input/key/RenderWindowInputHandler.kt
@@ -276,7 +276,7 @@ class RenderWindowInputHandler(
         return false
     }
 
-    fun draw(delta: Long) {
+    fun draw(delta: Double) {
         camera.draw(delta)
         leftClickHandler.draw(delta)
         rightClickHandler.draw(delta)
